cilium: ipsec, add ciliumIP to the IPIdentityPair in the ipcache
IPSec needs to know the cilium_host IP in order to add them to the
policy/state database. Additionally, in tunnel mode these are also used
as the outer IP addresses. Add a new field to the IPIdentityPair so we
distribute the cilium host IP address along with endpoint IPs.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>